Add tip states to a phylogenetic tree

add_tip_states() annotates the tips of a phylo object with state values.

Usage

add_tip_states(tree, tip_data)

Arguments

tree

An object of class phylo.

tip_data

Either a named vector or a data.frame. If a named vector, the elements of the vector must be the state of each tip and the names must be the labels of those tips. If a data.frame, there must be 2 columns: the first with the tip labels and the second with the tip states.

Value

An object of class phylo with annotated tip states.

Details

A tip.states attribute is added to tree based on the tip_data provided by the user. This function is used within the prepare_tree_for_saasi() function.

See also

Examples

demo_tree
#> $edge
#>       [,1] [,2]
#>  [1,]   14   15
#>  [2,]   15   16
#>  [3,]   16   17
#>  [4,]   17   18
#>  [5,]   18    3
#>  [6,]   18   19
#>  [7,]   19    8
#>  [8,]   19    9
#>  [9,]   17    2
#> [10,]   16   20
#> [11,]   20   21
#> [12,]   21   10
#> [13,]   21    7
#> [14,]   20    6
#> [15,]   15   22
#> [16,]   22   12
#> [17,]   22   11
#> [18,]   14   23
#> [19,]   23   24
#> [20,]   24   25
#> [21,]   25    5
#> [22,]   25   13
#> [23,]   24    4
#> [24,]   23    1
#> 
#> $Nnode
#> [1] 12
#> 
#> $tip.label
#>  [1] "sp2"  "sp4"  "sp8"  "sp9"  "sp10" "sp26" "sp27" "sp30" "sp31" "sp59"
#> [11] "sp64" "sp72" "sp75"
#> 
#> $node.label
#>  [1] "nd1"  "nd2"  "nd6"  "nd22" "nd27" "nd74" "nd32" "nd54" "nd7"  "nd8" 
#> [11] "nd36" "nd38"
#> 
#> $node.state
#>  nd1  nd2  nd6 nd22 nd27 nd74 nd32 nd54  nd7  nd8 nd36 nd38 
#>    1    1    2    2    2    2    2    2    1    1    1    1 
#> 
#> $edge.length
#>  [1] 0.16467463 0.60776479 0.06123598 0.13756928 0.19052413 0.43829041
#>  [7] 0.14919102 0.26797730 0.12387543 0.38247528 0.24214935 0.28936865
#> [13] 0.31255390 0.27670668 0.43629704 1.11293809 1.11550593 0.53064429
#> [19] 0.51423161 0.04706635 0.37366599 0.58693711 0.14126857 0.95898051
#> 
#> $orig
#>     idx         len parent     start state extinct split   sam idx2 parent2
#> 2     2 0.164674627      1 0.1508471     1   FALSE  TRUE FALSE  101     100
#> 3     3 0.095752083      1 0.1508471     1   FALSE  TRUE FALSE  102     100
#> 4     4 0.350260185      3 0.2465992     1   FALSE  TRUE FALSE  103     102
#> 5     5 0.221791997      3 0.2465992     1   FALSE  TRUE FALSE  104     102
#> 6     6 0.607764789      2 0.3155218     2   FALSE  TRUE FALSE  105     101
#> 7     7 0.436297042      2 0.3155218     1   FALSE  TRUE FALSE  106     101
#> 8     8 0.213100210      5 0.4683912     1   FALSE  TRUE FALSE  107     104
#> 9     9 0.178937370      5 0.4683912     1   FALSE  TRUE FALSE  108     104
#> 10   10 0.908357264      4 0.5968594     2   FALSE  TRUE FALSE  109     103
#> 11   11 0.208011176      4 0.5968594     1   FALSE  TRUE FALSE  110     103
#> 12   12 0.958787122      9 0.6473286     3   FALSE  TRUE FALSE  111     108
#> 13   13 0.285044132      9 0.6473286     1   FALSE  TRUE FALSE  112     108
#> 14   14 0.511629972      8 0.6814914     1   FALSE  TRUE FALSE  113     107
#> 15   15 0.187315709      8 0.6814914     1   FALSE  TRUE FALSE  114     107
#> 16   16 0.287068808      7 0.7518188     1   FALSE  TRUE FALSE  115     106
#> 17   17 0.233292244      7 0.7518188     1   FALSE  TRUE FALSE  116     106
#> 18   18 1.121845705     11 0.8048706     1   FALSE FALSE FALSE    1     110
#> 19   19 0.364675377     11 0.8048706     1   FALSE  TRUE FALSE  117     110
#> 20   20 0.011876720     15 0.8688071     1   FALSE  TRUE FALSE  118     114
#> 21   21 0.771664802     15 0.8688071     1   FALSE FALSE  TRUE    2     114
#> 22   22 0.404881489     20 0.8806839     1   FALSE  TRUE FALSE  119     118
#> 23   23 0.164910945     20 0.8806839     1   FALSE  TRUE FALSE  120     118
#> 24   24 0.061235983      6 0.9232865     2   FALSE  TRUE FALSE  121     105
#> 25   25 0.147439207      6 0.9232865     2   FALSE  TRUE FALSE  122     105
#> 26   26 0.994343566     13 0.9323727     1   FALSE FALSE FALSE    3     112
#> 27   27 0.718118368     13 0.9323727     3   FALSE  TRUE FALSE  123     112
#> 28   28 0.008634727     24 0.9845225     2   FALSE  TRUE FALSE  124     121
#> 29   29 0.123875431     24 0.9845225     2   FALSE FALSE  TRUE    4     121
#> 30   30 0.941605236     17 0.9851110     1   FALSE FALSE FALSE    5     116
#> 31   31 0.210691558     17 0.9851110     1   FALSE  TRUE FALSE  125     116
#> 33   33 0.128934554     28 0.9931573     2   FALSE  TRUE FALSE  126     124
#> 34   34 0.196808061     16 1.0388876     1   FALSE  TRUE FALSE  127     115
#> 35   35 0.175404690     16 1.0388876     1   FALSE  TRUE FALSE  128     115
#> 36   36 0.881121485     23 1.0455948     3   FALSE FALSE FALSE    7     120
#> 37   37 0.523725191     23 1.0455948     1   FALSE  TRUE FALSE  129     120
#> 38   38 0.296646520     25 1.0707258     2   FALSE  TRUE FALSE  130     122
#> 39   39 0.235036077     25 1.0707258     2   FALSE  TRUE FALSE  131     122
#> 40   40 0.190524130     33 1.1220918     2   FALSE FALSE  TRUE    8     126
#> 41   41 0.378207392     33 1.1220918     2   FALSE  TRUE FALSE  132     126
#> 42   42 0.088889776     19 1.1695460     1   FALSE  TRUE FALSE  133     117
#> 43   43 0.084562865     19 1.1695460     3   FALSE  TRUE FALSE  134     117
#> 44   44 0.002601639     14 1.1931214     1   FALSE  TRUE FALSE  135     113
#> 45   45 0.424346440     14 1.1931214     1   FALSE  TRUE FALSE  136     113
#> 46   46 0.047066352     44 1.1957230     1   FALSE  TRUE FALSE  137     135
#> 47   47 0.141268574     44 1.1957230     1   FALSE FALSE  TRUE    9     135
#> 48   48 0.146166009     31 1.1958026     1   FALSE  TRUE FALSE  138     125
#> 49   49 0.067324209     31 1.1958026     1   FALSE  TRUE FALSE  139     125
#> 50   50 0.544442729     35 1.2142923     1   FALSE  TRUE FALSE  140     128
#> 51   51 0.135830510     35 1.2142923     1   FALSE  TRUE FALSE  141     128
#> 52   52 0.054169952     34 1.2356957     1   FALSE  TRUE FALSE  142     127
#> 53   53 0.085441542     34 1.2356957     1   FALSE  TRUE FALSE  143     127
#> 54   54 0.373665988     46 1.2427894     3   FALSE FALSE  TRUE   10     137
#> 55   55 0.005947470     46 1.2427894     1   FALSE  TRUE FALSE  144     137
#> 56   56 0.062525080     55 1.2487369     1   FALSE  TRUE FALSE  145     144
#> 57   57 0.079893078     55 1.2487369     1   FALSE  TRUE FALSE  146     144
#> 58   58 0.672607463     43 1.2541088     3   FALSE FALSE FALSE   11     134
#> 59   59 0.314828098     43 1.2541088     3   FALSE  TRUE FALSE  147     134
#> 60   60 0.668280552     42 1.2584357     1   FALSE FALSE FALSE   12     133
#> 61   61 0.668280552     42 1.2584357     1   FALSE FALSE FALSE   13     133
#> 62   62 0.421234449     49 1.2631268     1   FALSE  TRUE FALSE  148     139
#> 63   63 0.069066626     49 1.2631268     1   FALSE  TRUE FALSE  149     139
#> 64   64 0.369898133     22 1.2855653     1   FALSE  TRUE FALSE  150     119
#> 65   65 0.641150941     22 1.2855653     1   FALSE FALSE FALSE   14     119
#> 66   66 0.280602490     52 1.2898656     1   FALSE  TRUE FALSE  151     142
#> 67   67 0.320210285     52 1.2898656     1   FALSE  TRUE FALSE  152     142
#> 68   68 0.242149352     39 1.3057618     2   FALSE  TRUE FALSE  153     131
#> 69   69 0.233226652     39 1.3057618     2   FALSE  TRUE FALSE  154     131
#> 70   70 0.570054335     56 1.3112619     1   FALSE  TRUE FALSE  155     145
#> 71   71 0.615454346     56 1.3112619     2   FALSE FALSE FALSE   15     145
#> 72   72 0.083934169     53 1.3211372     1   FALSE  TRUE FALSE  156     143
#> 73   73 0.593478377     53 1.3211372     1   FALSE  TRUE FALSE  157     143
#> 74   74 0.103403746     57 1.3286299     1   FALSE  TRUE FALSE  158     146
#> 75   75 0.432689222     57 1.3286299     1   FALSE  TRUE FALSE  159     146
#> 76   76 0.288974657     63 1.3321934     1   FALSE  TRUE FALSE  160     149
#> 77   77 0.187107326     63 1.3321934     1   FALSE  TRUE FALSE  161     149
#> 78   78 0.101989443     48 1.3419686     1   FALSE  TRUE FALSE  162     138
#> 79   79 0.358703541     48 1.3419686     1   FALSE  TRUE FALSE  163     138
#> 80   80 0.242264625     51 1.3501228     1   FALSE  TRUE FALSE  164     141
#> 81   81 0.576593472     51 1.3501228     3   FALSE FALSE FALSE   16     141
#> 82   82 0.043004976     38 1.3673723     2   FALSE  TRUE FALSE  165     130
#> 83   83 0.559344007     38 1.3673723     2   FALSE FALSE FALSE   17     130
#> 84   84 0.521644901     72 1.4050714     1   FALSE FALSE FALSE   18     156
#> 85   85 0.521644901     72 1.4050714     1   FALSE FALSE FALSE   19     156
#> 86   86 0.065834152     82 1.4103773     2   FALSE  TRUE FALSE  166     165
#> 87   87 0.261437974     82 1.4103773     2   FALSE  TRUE FALSE  167     165
#> 88   88 0.296346500     74 1.4320337     1   FALSE  TRUE FALSE  168     158
#> 89   89 0.097792699     74 1.4320337     1   FALSE  TRUE FALSE  169     158
#> 91   91 0.410609832     78 1.4439581     1   FALSE  TRUE FALSE  170     162
#> 92   92 0.293597865     86 1.4762114     2   FALSE  TRUE FALSE  171     166
#> 93   93 0.083985009     86 1.4762114     2   FALSE  TRUE FALSE  172     166
#> 94   94 0.426417078     41 1.5002992     2   FALSE FALSE FALSE   21     132
#> 95   95 0.060083014     41 1.5002992     2   FALSE  TRUE FALSE  173     132
#> 96   96 0.159140297     10 1.5052167     2   FALSE  TRUE FALSE  174     109
#> 97   97 0.421499617     10 1.5052167     2   FALSE FALSE FALSE   22     109
#> 98   98 0.407415518     77 1.5193008     1   FALSE FALSE FALSE   23     161
#> 99   99 0.407415518     77 1.5193008     1   FALSE FALSE FALSE   24     161
#> 100 100 0.092729426     89 1.5298264     1   FALSE  TRUE FALSE  175     169
#> 101 101 0.274753602     89 1.5298264     3   FALSE  TRUE FALSE  176     169
#> 102 102 0.387727798     69 1.5389885     2   FALSE FALSE FALSE   25     154
#> 103 103 0.043480033     69 1.5389885     2   FALSE FALSE  TRUE   26     154
#> 104 104 0.175579119     68 1.5479112     2   FALSE  TRUE FALSE  177     153
#> 105 105 0.312553900     68 1.5479112     3   FALSE FALSE  TRUE   27     153
#> 106 106 0.366519870     93 1.5601964     2   FALSE FALSE FALSE   28     172
#> 107 107 0.366519870     93 1.5601964     2   FALSE FALSE FALSE   29     172
#> 108 108 0.149191016     95 1.5603822     2   FALSE FALSE  TRUE   30     173
#> 109 109 0.267977300     95 1.5603822     3   FALSE FALSE  TRUE   31     173
#> 110 110 0.357779365     59 1.5689369     3   FALSE FALSE FALSE   32     147
#> 111 111 0.357779365     59 1.5689369     3   FALSE FALSE FALSE   33     147
#> 112 112 0.357396294     37 1.5693200     1   FALSE FALSE FALSE   34     129
#> 113 113 0.357396294     37 1.5693200     1   FALSE FALSE FALSE   35     129
#> 114 114 0.356248170     66 1.5704681     1   FALSE FALSE FALSE   36     151
#> 115 115 0.356248170     66 1.5704681     1   FALSE FALSE FALSE   37     151
#> 116 116 0.334328846     80 1.5923874     3   FALSE FALSE FALSE   38     164
#> 117 117 0.334328846     80 1.5923874     1   FALSE FALSE FALSE   39     164
#> 118 118 0.314805032     12 1.6061157     2   FALSE  TRUE FALSE  178     111
#> 119 119 0.222454996     12 1.6061157     3   FALSE  TRUE FALSE  179     111
#> 120 120 0.049710495     67 1.6100759     1   FALSE  TRUE FALSE  180     152
#> 121 121 0.316640375     67 1.6100759     1   FALSE FALSE FALSE   40     152
#> 122 122 0.309248446     45 1.6174678     1   FALSE FALSE FALSE   41     136
#> 123 123 0.309248446     45 1.6174678     1   FALSE FALSE FALSE   42     136
#> 124 124 0.102119268     76 1.6211681     1   FALSE  TRUE FALSE  181     160
#> 125 125 0.305548186     76 1.6211681     3   FALSE FALSE FALSE   43     160
#> 126 126 0.052236063    100 1.6225558     1   FALSE  TRUE FALSE  182     175
#> 127 127 0.304160477    100 1.6225558     3   FALSE FALSE FALSE   44     175
#> 128 128 0.276225198     27 1.6504911     3   FALSE FALSE FALSE   45     123
#> 129 129 0.016622859     27 1.6504911     3   FALSE  TRUE FALSE  183     123
#> 130 130 0.147642260     64 1.6554635     1   FALSE  TRUE FALSE  184     150
#> 131 131 0.271252808     64 1.6554635     1   FALSE FALSE FALSE   46     150
#> 132 132 0.266929879    120 1.6597864     1   FALSE FALSE FALSE   47     180
#> 133 133 0.266929879    120 1.6597864     1   FALSE FALSE FALSE   48     180
#> 134 134 0.191284415     96 1.6643570     2   FALSE  TRUE FALSE  185     174
#> 135 135 0.093227680     96 1.6643570     2   FALSE  TRUE FALSE  186     174
#> 136 136 0.259602339    129 1.6671139     3   FALSE FALSE FALSE   49     183
#> 137 137 0.259602339    129 1.6671139     1   FALSE FALSE FALSE   50     183
#> 139 139 0.254901057     87 1.6718152     2   FALSE FALSE FALSE   52     167
#> 140 140 0.054001298    126 1.6747919     1   FALSE  TRUE FALSE  187     182
#> 141 141 0.006521519    126 1.6747919     1   FALSE  TRUE FALSE  188     182
#> 142 142 0.245402895    141 1.6813134     1   FALSE FALSE FALSE   53     188
#> 143 143 0.245402895    141 1.6813134     1   FALSE FALSE FALSE   54     188
#> 144 144 0.242355020     62 1.6843613     1   FALSE FALSE FALSE   55     148
#> 146 146 0.226044128     79 1.7006722     1   FALSE FALSE FALSE   57     163
#> 147 147 0.197537764     79 1.7006722     1   FALSE  TRUE FALSE  189     163
#> 148 148 0.203428919    124 1.7232874     1   FALSE FALSE FALSE   58     181
#> 149 149 0.034087869    124 1.7232874     1   FALSE  TRUE FALSE  190     181
#> 150 150 0.113789534    104 1.7234903     2   FALSE FALSE  TRUE   59     177
#> 151 151 0.203225979    104 1.7234903     2   FALSE FALSE FALSE   60     177
#> 152 152 0.198336103     88 1.7283802     1   FALSE FALSE FALSE   61     168
#> 153 153 0.198336103     88 1.7283802     1   FALSE FALSE FALSE   62     168
#> 154 154 0.197923116    140 1.7287932     1   FALSE FALSE FALSE   63     187
#> 155 155 0.166374148    140 1.7287932     1   FALSE  TRUE FALSE  191     187
#> 156 156 0.014147720    149 1.7573752     1   FALSE  TRUE FALSE  192     190
#> 157 157 0.109949497    149 1.7573752     3   FALSE FALSE  TRUE   64     190
#> 158 158 0.169131640    135 1.7575846     3   FALSE FALSE FALSE   65     186
#> 159 159 0.169131640    135 1.7575846     2   FALSE FALSE FALSE   66     186
#> 160 160 0.157153882     50 1.7587350     1   FALSE  TRUE FALSE  193     140
#> 161 161 0.041193647     50 1.7587350     2   FALSE  TRUE FALSE  194     140
#> 162 162 0.121493122     75 1.7613192     1   FALSE  TRUE FALSE  195     159
#> 163 163 0.141347666     75 1.7613192     1   FALSE  TRUE FALSE  196     159
#> 164 164 0.156907014     92 1.7698093     2   FALSE FALSE FALSE   67     171
#> 165 165 0.156907014     92 1.7698093     2   FALSE FALSE FALSE   68     171
#> 166 166 0.155193329    156 1.7715230     1   FALSE FALSE FALSE   69     192
#> 167 167 0.155193329    156 1.7715230     1   FALSE FALSE FALSE   70     192
#> 168 168 0.126787606    161 1.7999287     2   FALSE FALSE FALSE   71     194
#> 169 169 0.064828216    161 1.7999287     2   FALSE FALSE  TRUE   72     194
#> 170 170 0.123610548    130 1.8031057     1   FALSE FALSE FALSE   73     184
#> 171 171 0.123610548    130 1.8031057     1   FALSE FALSE FALSE   74     184
#> 172 172 0.058918845    101 1.8045800     3   FALSE  TRUE FALSE  197     176
#> 173 173 0.025146519    101 1.8045800     3   FALSE FALSE  TRUE   75     176
#> 174 174 0.098145581    119 1.8285707     3   FALSE FALSE FALSE   76     179
#> 176 176 0.072148394     91 1.8545679     1   FALSE FALSE FALSE   78     170
#> 177 177 0.072148394     91 1.8545679     1   FALSE FALSE FALSE   79     170
#> 178 178 0.071074905    134 1.8556414     2   FALSE FALSE FALSE   80     185
#> 179 179 0.071074905    134 1.8556414     2   FALSE FALSE FALSE   81     185
#> 180 180 0.063217456    172 1.8634988     3   FALSE FALSE FALSE   82     197
#> 181 181 0.063217456    172 1.8634988     3   FALSE FALSE FALSE   83     197
#> 182 182 0.045400011     70 1.8813163     1   FALSE FALSE FALSE   84     155
#> 183 183 0.045400011     70 1.8813163     1   FALSE FALSE FALSE   85     155
#> 184 184 0.043904005    162 1.8828123     1   FALSE FALSE FALSE   86     195
#> 185 185 0.043904005    162 1.8828123     1   FALSE FALSE FALSE   87     195
#> 186 186 0.031548968    155 1.8951673     1   FALSE FALSE FALSE   88     191
#> 187 187 0.031548968    155 1.8951673     1   FALSE FALSE FALSE   89     191
#> 188 188 0.028506365    147 1.8982099     1   FALSE FALSE FALSE   90     189
#> 190 190 0.024049461    163 1.9026668     1   FALSE FALSE FALSE   92     196
#> 191 191 0.024049461    163 1.9026668     1   FALSE FALSE FALSE   93     196
#> 192 192 0.012100693     73 1.9146156     1   FALSE FALSE FALSE   94     157
#> 193 193 0.012100693     73 1.9146156     1   FALSE FALSE FALSE   95     157
#> 194 194 0.010827371    160 1.9158889     1   FALSE FALSE FALSE   96     193
#> 195 195 0.010827371    160 1.9158889     1   FALSE FALSE FALSE   97     193
#> 196 196 0.005795544    118 1.9209207     2   FALSE FALSE FALSE   98     178
#> 197 197 0.005795544    118 1.9209207     2   FALSE FALSE FALSE   99     178
#>     name name2
#> 2   <NA>   nd2
#> 3   <NA>   nd3
#> 4   <NA>   nd4
#> 5   <NA>   nd5
#> 6   <NA>   nd6
#> 7   <NA>   nd7
#> 8   <NA>   nd8
#> 9   <NA>   nd9
#> 10  <NA>  nd10
#> 11  <NA>  nd11
#> 12  <NA>  nd12
#> 13  <NA>  nd13
#> 14  <NA>  nd14
#> 15  <NA>  nd15
#> 16  <NA>  nd16
#> 17  <NA>  nd17
#> 18   sp1   sp1
#> 19  <NA>  nd18
#> 20  <NA>  nd19
#> 21   sp2   sp2
#> 22  <NA>  nd20
#> 23  <NA>  nd21
#> 24  <NA>  nd22
#> 25  <NA>  nd23
#> 26   sp3   sp3
#> 27  <NA>  nd24
#> 28  <NA>  nd25
#> 29   sp4   sp4
#> 30   sp5   sp5
#> 31  <NA>  nd26
#> 33  <NA>  nd27
#> 34  <NA>  nd28
#> 35  <NA>  nd29
#> 36   sp7   sp7
#> 37  <NA>  nd30
#> 38  <NA>  nd31
#> 39  <NA>  nd32
#> 40   sp8   sp8
#> 41  <NA>  nd33
#> 42  <NA>  nd34
#> 43  <NA>  nd35
#> 44  <NA>  nd36
#> 45  <NA>  nd37
#> 46  <NA>  nd38
#> 47   sp9   sp9
#> 48  <NA>  nd39
#> 49  <NA>  nd40
#> 50  <NA>  nd41
#> 51  <NA>  nd42
#> 52  <NA>  nd43
#> 53  <NA>  nd44
#> 54  sp10  sp10
#> 55  <NA>  nd45
#> 56  <NA>  nd46
#> 57  <NA>  nd47
#> 58  sp11  sp11
#> 59  <NA>  nd48
#> 60  sp12  sp12
#> 61  sp13  sp13
#> 62  <NA>  nd49
#> 63  <NA>  nd50
#> 64  <NA>  nd51
#> 65  sp14  sp14
#> 66  <NA>  nd52
#> 67  <NA>  nd53
#> 68  <NA>  nd54
#> 69  <NA>  nd55
#> 70  <NA>  nd56
#> 71  sp15  sp15
#> 72  <NA>  nd57
#> 73  <NA>  nd58
#> 74  <NA>  nd59
#> 75  <NA>  nd60
#> 76  <NA>  nd61
#> 77  <NA>  nd62
#> 78  <NA>  nd63
#> 79  <NA>  nd64
#> 80  <NA>  nd65
#> 81  sp16  sp16
#> 82  <NA>  nd66
#> 83  sp17  sp17
#> 84  sp18  sp18
#> 85  sp19  sp19
#> 86  <NA>  nd67
#> 87  <NA>  nd68
#> 88  <NA>  nd69
#> 89  <NA>  nd70
#> 91  <NA>  nd71
#> 92  <NA>  nd72
#> 93  <NA>  nd73
#> 94  sp21  sp21
#> 95  <NA>  nd74
#> 96  <NA>  nd75
#> 97  sp22  sp22
#> 98  sp23  sp23
#> 99  sp24  sp24
#> 100 <NA>  nd76
#> 101 <NA>  nd77
#> 102 sp25  sp25
#> 103 sp26  sp26
#> 104 <NA>  nd78
#> 105 sp27  sp27
#> 106 sp28  sp28
#> 107 sp29  sp29
#> 108 sp30  sp30
#> 109 sp31  sp31
#> 110 sp32  sp32
#> 111 sp33  sp33
#> 112 sp34  sp34
#> 113 sp35  sp35
#> 114 sp36  sp36
#> 115 sp37  sp37
#> 116 sp38  sp38
#> 117 sp39  sp39
#> 118 <NA>  nd79
#> 119 <NA>  nd80
#> 120 <NA>  nd81
#> 121 sp40  sp40
#> 122 sp41  sp41
#> 123 sp42  sp42
#> 124 <NA>  nd82
#> 125 sp43  sp43
#> 126 <NA>  nd83
#> 127 sp44  sp44
#> 128 sp45  sp45
#> 129 <NA>  nd84
#> 130 <NA>  nd85
#> 131 sp46  sp46
#> 132 sp47  sp47
#> 133 sp48  sp48
#> 134 <NA>  nd86
#> 135 <NA>  nd87
#> 136 sp49  sp49
#> 137 sp50  sp50
#> 139 sp52  sp52
#> 140 <NA>  nd88
#> 141 <NA>  nd89
#> 142 sp53  sp53
#> 143 sp54  sp54
#> 144 sp55  sp55
#> 146 sp57  sp57
#> 147 <NA>  nd90
#> 148 sp58  sp58
#> 149 <NA>  nd91
#> 150 sp59  sp59
#> 151 sp60  sp60
#> 152 sp61  sp61
#> 153 sp62  sp62
#> 154 sp63  sp63
#> 155 <NA>  nd92
#> 156 <NA>  nd93
#> 157 sp64  sp64
#> 158 sp65  sp65
#> 159 sp66  sp66
#> 160 <NA>  nd94
#> 161 <NA>  nd95
#> 162 <NA>  nd96
#> 163 <NA>  nd97
#> 164 sp67  sp67
#> 165 sp68  sp68
#> 166 sp69  sp69
#> 167 sp70  sp70
#> 168 sp71  sp71
#> 169 sp72  sp72
#> 170 sp73  sp73
#> 171 sp74  sp74
#> 172 <NA>  nd98
#> 173 sp75  sp75
#> 174 sp76  sp76
#> 176 sp78  sp78
#> 177 sp79  sp79
#> 178 sp80  sp80
#> 179 sp81  sp81
#> 180 sp82  sp82
#> 181 sp83  sp83
#> 182 sp84  sp84
#> 183 sp85  sp85
#> 184 sp86  sp86
#> 185 sp87  sp87
#> 186 sp88  sp88
#> 187 sp89  sp89
#> 188 sp90  sp90
#> 190 sp92  sp92
#> 191 sp93  sp93
#> 192 sp94  sp94
#> 193 sp95  sp95
#> 194 sp96  sp96
#> 195 sp97  sp97
#> 196 sp98  sp98
#> 197 sp99  sp99
#> 
#> $hist
#>    idx         t from to        x0         tc idx2 name2
#> 13  36 1.7712115    1  3 1.0455948 0.72561669    6   sp7
#> 7   54 1.4714900    1  3 1.2427894 0.22870062    9  sp10
#> 11  71 1.7523062    1  2 1.3112619 0.44104431   14  sp15
#> 6   81 1.3637622    1  3 1.3501228 0.01363937   15  sp16
#> 12 105 1.7655396    2  3 1.5479112 0.21762839   25  sp27
#> 9  109 1.6938940    2  3 1.5603822 0.13351183   29  sp31
#> 17 116 1.8579434    1  3 1.5923874 0.26555601   36  sp38
#> 20 125 1.9110279    1  3 1.6211681 0.28985983   41  sp43
#> 15 127 1.8226049    1  3 1.6225558 0.20004911   42  sp44
#> 10 137 1.7423445    3  1 1.6671139 0.07523055   48  sp50
#> 16 157 1.8506348    1  3 1.7573752 0.09325955   60  sp64
#> 19 158 1.9005004    2  3 1.7575846 0.14291580   61  sp65
#> 2    6 0.7937786    1  2 0.3155218 0.47825687   96   nd6
#> 14 161 1.7896296    1  2 1.7587350 0.03089453  124  nd95
#> 1   10 0.7664957    1  2 0.5968594 0.16963631  141  nd10
#> 5   43 1.2514209    1  3 1.1695460 0.08187492  148  nd35
#> 8  101 1.6656791    1  3 1.5298264 0.13585273  167  nd77
#> 4   12 1.1409286    1  3 0.6473286 0.49360002  181  nd12
#> 18 118 1.8697629    3  2 1.6061157 0.26364715  182  nd79
#> 3   27 1.1014641    1  3 0.9323727 0.16909142  184  nd24
#> 
#> $edge.state
#>   [1] 1 2 2 2 2 2 2 2 2 2 2 3 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 3 2 2 2 1
#>  [38] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 2 2 2 1 1 3 1 3 1 1 1 1 1 1
#>  [75] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 3 3 1 1 1 1 1 2 2 2 2 2 2 3 2 2 1
#> [112] 1 1 1 1 1 3 3 3 3 3 1 1 1 1 1 3 1 1 1 1 1 2 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [149] 1 3 3 3 3 3 3 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 3 1 1 1 1 1 3 2 2
#> [186] 2 3 3 3 1 1 3 3 3 3 1
#> 
#> attr(,"class")
#> [1] "phylo"
#> attr(,"order")
#> [1] "cladewise"
head(demo_metadata)
#>   node state
#> 1  sp2     1
#> 2  sp4     2
#> 3  sp8     2
#> 4  sp9     1
#> 5 sp10     3
#> 6 sp26     2

add_tip_states(demo_tree, demo_metadata)
#> $edge
#>       [,1] [,2]
#>  [1,]   14   15
#>  [2,]   15   16
#>  [3,]   16   17
#>  [4,]   17   18
#>  [5,]   18    3
#>  [6,]   18   19
#>  [7,]   19    8
#>  [8,]   19    9
#>  [9,]   17    2
#> [10,]   16   20
#> [11,]   20   21
#> [12,]   21   10
#> [13,]   21    7
#> [14,]   20    6
#> [15,]   15   22
#> [16,]   22   12
#> [17,]   22   11
#> [18,]   14   23
#> [19,]   23   24
#> [20,]   24   25
#> [21,]   25    5
#> [22,]   25   13
#> [23,]   24    4
#> [24,]   23    1
#> 
#> $Nnode
#> [1] 12
#> 
#> $tip.label
#>  [1] "sp2"  "sp4"  "sp8"  "sp9"  "sp10" "sp26" "sp27" "sp30" "sp31" "sp59"
#> [11] "sp64" "sp72" "sp75"
#> 
#> $node.label
#>  [1] "nd1"  "nd2"  "nd6"  "nd22" "nd27" "nd74" "nd32" "nd54" "nd7"  "nd8" 
#> [11] "nd36" "nd38"
#> 
#> $node.state
#>  nd1  nd2  nd6 nd22 nd27 nd74 nd32 nd54  nd7  nd8 nd36 nd38 
#>    1    1    2    2    2    2    2    2    1    1    1    1 
#> 
#> $edge.length
#>  [1] 0.16467463 0.60776479 0.06123598 0.13756928 0.19052413 0.43829041
#>  [7] 0.14919102 0.26797730 0.12387543 0.38247528 0.24214935 0.28936865
#> [13] 0.31255390 0.27670668 0.43629704 1.11293809 1.11550593 0.53064429
#> [19] 0.51423161 0.04706635 0.37366599 0.58693711 0.14126857 0.95898051
#> 
#> $orig
#>     idx         len parent     start state extinct split   sam idx2 parent2
#> 2     2 0.164674627      1 0.1508471     1   FALSE  TRUE FALSE  101     100
#> 3     3 0.095752083      1 0.1508471     1   FALSE  TRUE FALSE  102     100
#> 4     4 0.350260185      3 0.2465992     1   FALSE  TRUE FALSE  103     102
#> 5     5 0.221791997      3 0.2465992     1   FALSE  TRUE FALSE  104     102
#> 6     6 0.607764789      2 0.3155218     2   FALSE  TRUE FALSE  105     101
#> 7     7 0.436297042      2 0.3155218     1   FALSE  TRUE FALSE  106     101
#> 8     8 0.213100210      5 0.4683912     1   FALSE  TRUE FALSE  107     104
#> 9     9 0.178937370      5 0.4683912     1   FALSE  TRUE FALSE  108     104
#> 10   10 0.908357264      4 0.5968594     2   FALSE  TRUE FALSE  109     103
#> 11   11 0.208011176      4 0.5968594     1   FALSE  TRUE FALSE  110     103
#> 12   12 0.958787122      9 0.6473286     3   FALSE  TRUE FALSE  111     108
#> 13   13 0.285044132      9 0.6473286     1   FALSE  TRUE FALSE  112     108
#> 14   14 0.511629972      8 0.6814914     1   FALSE  TRUE FALSE  113     107
#> 15   15 0.187315709      8 0.6814914     1   FALSE  TRUE FALSE  114     107
#> 16   16 0.287068808      7 0.7518188     1   FALSE  TRUE FALSE  115     106
#> 17   17 0.233292244      7 0.7518188     1   FALSE  TRUE FALSE  116     106
#> 18   18 1.121845705     11 0.8048706     1   FALSE FALSE FALSE    1     110
#> 19   19 0.364675377     11 0.8048706     1   FALSE  TRUE FALSE  117     110
#> 20   20 0.011876720     15 0.8688071     1   FALSE  TRUE FALSE  118     114
#> 21   21 0.771664802     15 0.8688071     1   FALSE FALSE  TRUE    2     114
#> 22   22 0.404881489     20 0.8806839     1   FALSE  TRUE FALSE  119     118
#> 23   23 0.164910945     20 0.8806839     1   FALSE  TRUE FALSE  120     118
#> 24   24 0.061235983      6 0.9232865     2   FALSE  TRUE FALSE  121     105
#> 25   25 0.147439207      6 0.9232865     2   FALSE  TRUE FALSE  122     105
#> 26   26 0.994343566     13 0.9323727     1   FALSE FALSE FALSE    3     112
#> 27   27 0.718118368     13 0.9323727     3   FALSE  TRUE FALSE  123     112
#> 28   28 0.008634727     24 0.9845225     2   FALSE  TRUE FALSE  124     121
#> 29   29 0.123875431     24 0.9845225     2   FALSE FALSE  TRUE    4     121
#> 30   30 0.941605236     17 0.9851110     1   FALSE FALSE FALSE    5     116
#> 31   31 0.210691558     17 0.9851110     1   FALSE  TRUE FALSE  125     116
#> 33   33 0.128934554     28 0.9931573     2   FALSE  TRUE FALSE  126     124
#> 34   34 0.196808061     16 1.0388876     1   FALSE  TRUE FALSE  127     115
#> 35   35 0.175404690     16 1.0388876     1   FALSE  TRUE FALSE  128     115
#> 36   36 0.881121485     23 1.0455948     3   FALSE FALSE FALSE    7     120
#> 37   37 0.523725191     23 1.0455948     1   FALSE  TRUE FALSE  129     120
#> 38   38 0.296646520     25 1.0707258     2   FALSE  TRUE FALSE  130     122
#> 39   39 0.235036077     25 1.0707258     2   FALSE  TRUE FALSE  131     122
#> 40   40 0.190524130     33 1.1220918     2   FALSE FALSE  TRUE    8     126
#> 41   41 0.378207392     33 1.1220918     2   FALSE  TRUE FALSE  132     126
#> 42   42 0.088889776     19 1.1695460     1   FALSE  TRUE FALSE  133     117
#> 43   43 0.084562865     19 1.1695460     3   FALSE  TRUE FALSE  134     117
#> 44   44 0.002601639     14 1.1931214     1   FALSE  TRUE FALSE  135     113
#> 45   45 0.424346440     14 1.1931214     1   FALSE  TRUE FALSE  136     113
#> 46   46 0.047066352     44 1.1957230     1   FALSE  TRUE FALSE  137     135
#> 47   47 0.141268574     44 1.1957230     1   FALSE FALSE  TRUE    9     135
#> 48   48 0.146166009     31 1.1958026     1   FALSE  TRUE FALSE  138     125
#> 49   49 0.067324209     31 1.1958026     1   FALSE  TRUE FALSE  139     125
#> 50   50 0.544442729     35 1.2142923     1   FALSE  TRUE FALSE  140     128
#> 51   51 0.135830510     35 1.2142923     1   FALSE  TRUE FALSE  141     128
#> 52   52 0.054169952     34 1.2356957     1   FALSE  TRUE FALSE  142     127
#> 53   53 0.085441542     34 1.2356957     1   FALSE  TRUE FALSE  143     127
#> 54   54 0.373665988     46 1.2427894     3   FALSE FALSE  TRUE   10     137
#> 55   55 0.005947470     46 1.2427894     1   FALSE  TRUE FALSE  144     137
#> 56   56 0.062525080     55 1.2487369     1   FALSE  TRUE FALSE  145     144
#> 57   57 0.079893078     55 1.2487369     1   FALSE  TRUE FALSE  146     144
#> 58   58 0.672607463     43 1.2541088     3   FALSE FALSE FALSE   11     134
#> 59   59 0.314828098     43 1.2541088     3   FALSE  TRUE FALSE  147     134
#> 60   60 0.668280552     42 1.2584357     1   FALSE FALSE FALSE   12     133
#> 61   61 0.668280552     42 1.2584357     1   FALSE FALSE FALSE   13     133
#> 62   62 0.421234449     49 1.2631268     1   FALSE  TRUE FALSE  148     139
#> 63   63 0.069066626     49 1.2631268     1   FALSE  TRUE FALSE  149     139
#> 64   64 0.369898133     22 1.2855653     1   FALSE  TRUE FALSE  150     119
#> 65   65 0.641150941     22 1.2855653     1   FALSE FALSE FALSE   14     119
#> 66   66 0.280602490     52 1.2898656     1   FALSE  TRUE FALSE  151     142
#> 67   67 0.320210285     52 1.2898656     1   FALSE  TRUE FALSE  152     142
#> 68   68 0.242149352     39 1.3057618     2   FALSE  TRUE FALSE  153     131
#> 69   69 0.233226652     39 1.3057618     2   FALSE  TRUE FALSE  154     131
#> 70   70 0.570054335     56 1.3112619     1   FALSE  TRUE FALSE  155     145
#> 71   71 0.615454346     56 1.3112619     2   FALSE FALSE FALSE   15     145
#> 72   72 0.083934169     53 1.3211372     1   FALSE  TRUE FALSE  156     143
#> 73   73 0.593478377     53 1.3211372     1   FALSE  TRUE FALSE  157     143
#> 74   74 0.103403746     57 1.3286299     1   FALSE  TRUE FALSE  158     146
#> 75   75 0.432689222     57 1.3286299     1   FALSE  TRUE FALSE  159     146
#> 76   76 0.288974657     63 1.3321934     1   FALSE  TRUE FALSE  160     149
#> 77   77 0.187107326     63 1.3321934     1   FALSE  TRUE FALSE  161     149
#> 78   78 0.101989443     48 1.3419686     1   FALSE  TRUE FALSE  162     138
#> 79   79 0.358703541     48 1.3419686     1   FALSE  TRUE FALSE  163     138
#> 80   80 0.242264625     51 1.3501228     1   FALSE  TRUE FALSE  164     141
#> 81   81 0.576593472     51 1.3501228     3   FALSE FALSE FALSE   16     141
#> 82   82 0.043004976     38 1.3673723     2   FALSE  TRUE FALSE  165     130
#> 83   83 0.559344007     38 1.3673723     2   FALSE FALSE FALSE   17     130
#> 84   84 0.521644901     72 1.4050714     1   FALSE FALSE FALSE   18     156
#> 85   85 0.521644901     72 1.4050714     1   FALSE FALSE FALSE   19     156
#> 86   86 0.065834152     82 1.4103773     2   FALSE  TRUE FALSE  166     165
#> 87   87 0.261437974     82 1.4103773     2   FALSE  TRUE FALSE  167     165
#> 88   88 0.296346500     74 1.4320337     1   FALSE  TRUE FALSE  168     158
#> 89   89 0.097792699     74 1.4320337     1   FALSE  TRUE FALSE  169     158
#> 91   91 0.410609832     78 1.4439581     1   FALSE  TRUE FALSE  170     162
#> 92   92 0.293597865     86 1.4762114     2   FALSE  TRUE FALSE  171     166
#> 93   93 0.083985009     86 1.4762114     2   FALSE  TRUE FALSE  172     166
#> 94   94 0.426417078     41 1.5002992     2   FALSE FALSE FALSE   21     132
#> 95   95 0.060083014     41 1.5002992     2   FALSE  TRUE FALSE  173     132
#> 96   96 0.159140297     10 1.5052167     2   FALSE  TRUE FALSE  174     109
#> 97   97 0.421499617     10 1.5052167     2   FALSE FALSE FALSE   22     109
#> 98   98 0.407415518     77 1.5193008     1   FALSE FALSE FALSE   23     161
#> 99   99 0.407415518     77 1.5193008     1   FALSE FALSE FALSE   24     161
#> 100 100 0.092729426     89 1.5298264     1   FALSE  TRUE FALSE  175     169
#> 101 101 0.274753602     89 1.5298264     3   FALSE  TRUE FALSE  176     169
#> 102 102 0.387727798     69 1.5389885     2   FALSE FALSE FALSE   25     154
#> 103 103 0.043480033     69 1.5389885     2   FALSE FALSE  TRUE   26     154
#> 104 104 0.175579119     68 1.5479112     2   FALSE  TRUE FALSE  177     153
#> 105 105 0.312553900     68 1.5479112     3   FALSE FALSE  TRUE   27     153
#> 106 106 0.366519870     93 1.5601964     2   FALSE FALSE FALSE   28     172
#> 107 107 0.366519870     93 1.5601964     2   FALSE FALSE FALSE   29     172
#> 108 108 0.149191016     95 1.5603822     2   FALSE FALSE  TRUE   30     173
#> 109 109 0.267977300     95 1.5603822     3   FALSE FALSE  TRUE   31     173
#> 110 110 0.357779365     59 1.5689369     3   FALSE FALSE FALSE   32     147
#> 111 111 0.357779365     59 1.5689369     3   FALSE FALSE FALSE   33     147
#> 112 112 0.357396294     37 1.5693200     1   FALSE FALSE FALSE   34     129
#> 113 113 0.357396294     37 1.5693200     1   FALSE FALSE FALSE   35     129
#> 114 114 0.356248170     66 1.5704681     1   FALSE FALSE FALSE   36     151
#> 115 115 0.356248170     66 1.5704681     1   FALSE FALSE FALSE   37     151
#> 116 116 0.334328846     80 1.5923874     3   FALSE FALSE FALSE   38     164
#> 117 117 0.334328846     80 1.5923874     1   FALSE FALSE FALSE   39     164
#> 118 118 0.314805032     12 1.6061157     2   FALSE  TRUE FALSE  178     111
#> 119 119 0.222454996     12 1.6061157     3   FALSE  TRUE FALSE  179     111
#> 120 120 0.049710495     67 1.6100759     1   FALSE  TRUE FALSE  180     152
#> 121 121 0.316640375     67 1.6100759     1   FALSE FALSE FALSE   40     152
#> 122 122 0.309248446     45 1.6174678     1   FALSE FALSE FALSE   41     136
#> 123 123 0.309248446     45 1.6174678     1   FALSE FALSE FALSE   42     136
#> 124 124 0.102119268     76 1.6211681     1   FALSE  TRUE FALSE  181     160
#> 125 125 0.305548186     76 1.6211681     3   FALSE FALSE FALSE   43     160
#> 126 126 0.052236063    100 1.6225558     1   FALSE  TRUE FALSE  182     175
#> 127 127 0.304160477    100 1.6225558     3   FALSE FALSE FALSE   44     175
#> 128 128 0.276225198     27 1.6504911     3   FALSE FALSE FALSE   45     123
#> 129 129 0.016622859     27 1.6504911     3   FALSE  TRUE FALSE  183     123
#> 130 130 0.147642260     64 1.6554635     1   FALSE  TRUE FALSE  184     150
#> 131 131 0.271252808     64 1.6554635     1   FALSE FALSE FALSE   46     150
#> 132 132 0.266929879    120 1.6597864     1   FALSE FALSE FALSE   47     180
#> 133 133 0.266929879    120 1.6597864     1   FALSE FALSE FALSE   48     180
#> 134 134 0.191284415     96 1.6643570     2   FALSE  TRUE FALSE  185     174
#> 135 135 0.093227680     96 1.6643570     2   FALSE  TRUE FALSE  186     174
#> 136 136 0.259602339    129 1.6671139     3   FALSE FALSE FALSE   49     183
#> 137 137 0.259602339    129 1.6671139     1   FALSE FALSE FALSE   50     183
#> 139 139 0.254901057     87 1.6718152     2   FALSE FALSE FALSE   52     167
#> 140 140 0.054001298    126 1.6747919     1   FALSE  TRUE FALSE  187     182
#> 141 141 0.006521519    126 1.6747919     1   FALSE  TRUE FALSE  188     182
#> 142 142 0.245402895    141 1.6813134     1   FALSE FALSE FALSE   53     188
#> 143 143 0.245402895    141 1.6813134     1   FALSE FALSE FALSE   54     188
#> 144 144 0.242355020     62 1.6843613     1   FALSE FALSE FALSE   55     148
#> 146 146 0.226044128     79 1.7006722     1   FALSE FALSE FALSE   57     163
#> 147 147 0.197537764     79 1.7006722     1   FALSE  TRUE FALSE  189     163
#> 148 148 0.203428919    124 1.7232874     1   FALSE FALSE FALSE   58     181
#> 149 149 0.034087869    124 1.7232874     1   FALSE  TRUE FALSE  190     181
#> 150 150 0.113789534    104 1.7234903     2   FALSE FALSE  TRUE   59     177
#> 151 151 0.203225979    104 1.7234903     2   FALSE FALSE FALSE   60     177
#> 152 152 0.198336103     88 1.7283802     1   FALSE FALSE FALSE   61     168
#> 153 153 0.198336103     88 1.7283802     1   FALSE FALSE FALSE   62     168
#> 154 154 0.197923116    140 1.7287932     1   FALSE FALSE FALSE   63     187
#> 155 155 0.166374148    140 1.7287932     1   FALSE  TRUE FALSE  191     187
#> 156 156 0.014147720    149 1.7573752     1   FALSE  TRUE FALSE  192     190
#> 157 157 0.109949497    149 1.7573752     3   FALSE FALSE  TRUE   64     190
#> 158 158 0.169131640    135 1.7575846     3   FALSE FALSE FALSE   65     186
#> 159 159 0.169131640    135 1.7575846     2   FALSE FALSE FALSE   66     186
#> 160 160 0.157153882     50 1.7587350     1   FALSE  TRUE FALSE  193     140
#> 161 161 0.041193647     50 1.7587350     2   FALSE  TRUE FALSE  194     140
#> 162 162 0.121493122     75 1.7613192     1   FALSE  TRUE FALSE  195     159
#> 163 163 0.141347666     75 1.7613192     1   FALSE  TRUE FALSE  196     159
#> 164 164 0.156907014     92 1.7698093     2   FALSE FALSE FALSE   67     171
#> 165 165 0.156907014     92 1.7698093     2   FALSE FALSE FALSE   68     171
#> 166 166 0.155193329    156 1.7715230     1   FALSE FALSE FALSE   69     192
#> 167 167 0.155193329    156 1.7715230     1   FALSE FALSE FALSE   70     192
#> 168 168 0.126787606    161 1.7999287     2   FALSE FALSE FALSE   71     194
#> 169 169 0.064828216    161 1.7999287     2   FALSE FALSE  TRUE   72     194
#> 170 170 0.123610548    130 1.8031057     1   FALSE FALSE FALSE   73     184
#> 171 171 0.123610548    130 1.8031057     1   FALSE FALSE FALSE   74     184
#> 172 172 0.058918845    101 1.8045800     3   FALSE  TRUE FALSE  197     176
#> 173 173 0.025146519    101 1.8045800     3   FALSE FALSE  TRUE   75     176
#> 174 174 0.098145581    119 1.8285707     3   FALSE FALSE FALSE   76     179
#> 176 176 0.072148394     91 1.8545679     1   FALSE FALSE FALSE   78     170
#> 177 177 0.072148394     91 1.8545679     1   FALSE FALSE FALSE   79     170
#> 178 178 0.071074905    134 1.8556414     2   FALSE FALSE FALSE   80     185
#> 179 179 0.071074905    134 1.8556414     2   FALSE FALSE FALSE   81     185
#> 180 180 0.063217456    172 1.8634988     3   FALSE FALSE FALSE   82     197
#> 181 181 0.063217456    172 1.8634988     3   FALSE FALSE FALSE   83     197
#> 182 182 0.045400011     70 1.8813163     1   FALSE FALSE FALSE   84     155
#> 183 183 0.045400011     70 1.8813163     1   FALSE FALSE FALSE   85     155
#> 184 184 0.043904005    162 1.8828123     1   FALSE FALSE FALSE   86     195
#> 185 185 0.043904005    162 1.8828123     1   FALSE FALSE FALSE   87     195
#> 186 186 0.031548968    155 1.8951673     1   FALSE FALSE FALSE   88     191
#> 187 187 0.031548968    155 1.8951673     1   FALSE FALSE FALSE   89     191
#> 188 188 0.028506365    147 1.8982099     1   FALSE FALSE FALSE   90     189
#> 190 190 0.024049461    163 1.9026668     1   FALSE FALSE FALSE   92     196
#> 191 191 0.024049461    163 1.9026668     1   FALSE FALSE FALSE   93     196
#> 192 192 0.012100693     73 1.9146156     1   FALSE FALSE FALSE   94     157
#> 193 193 0.012100693     73 1.9146156     1   FALSE FALSE FALSE   95     157
#> 194 194 0.010827371    160 1.9158889     1   FALSE FALSE FALSE   96     193
#> 195 195 0.010827371    160 1.9158889     1   FALSE FALSE FALSE   97     193
#> 196 196 0.005795544    118 1.9209207     2   FALSE FALSE FALSE   98     178
#> 197 197 0.005795544    118 1.9209207     2   FALSE FALSE FALSE   99     178
#>     name name2
#> 2   <NA>   nd2
#> 3   <NA>   nd3
#> 4   <NA>   nd4
#> 5   <NA>   nd5
#> 6   <NA>   nd6
#> 7   <NA>   nd7
#> 8   <NA>   nd8
#> 9   <NA>   nd9
#> 10  <NA>  nd10
#> 11  <NA>  nd11
#> 12  <NA>  nd12
#> 13  <NA>  nd13
#> 14  <NA>  nd14
#> 15  <NA>  nd15
#> 16  <NA>  nd16
#> 17  <NA>  nd17
#> 18   sp1   sp1
#> 19  <NA>  nd18
#> 20  <NA>  nd19
#> 21   sp2   sp2
#> 22  <NA>  nd20
#> 23  <NA>  nd21
#> 24  <NA>  nd22
#> 25  <NA>  nd23
#> 26   sp3   sp3
#> 27  <NA>  nd24
#> 28  <NA>  nd25
#> 29   sp4   sp4
#> 30   sp5   sp5
#> 31  <NA>  nd26
#> 33  <NA>  nd27
#> 34  <NA>  nd28
#> 35  <NA>  nd29
#> 36   sp7   sp7
#> 37  <NA>  nd30
#> 38  <NA>  nd31
#> 39  <NA>  nd32
#> 40   sp8   sp8
#> 41  <NA>  nd33
#> 42  <NA>  nd34
#> 43  <NA>  nd35
#> 44  <NA>  nd36
#> 45  <NA>  nd37
#> 46  <NA>  nd38
#> 47   sp9   sp9
#> 48  <NA>  nd39
#> 49  <NA>  nd40
#> 50  <NA>  nd41
#> 51  <NA>  nd42
#> 52  <NA>  nd43
#> 53  <NA>  nd44
#> 54  sp10  sp10
#> 55  <NA>  nd45
#> 56  <NA>  nd46
#> 57  <NA>  nd47
#> 58  sp11  sp11
#> 59  <NA>  nd48
#> 60  sp12  sp12
#> 61  sp13  sp13
#> 62  <NA>  nd49
#> 63  <NA>  nd50
#> 64  <NA>  nd51
#> 65  sp14  sp14
#> 66  <NA>  nd52
#> 67  <NA>  nd53
#> 68  <NA>  nd54
#> 69  <NA>  nd55
#> 70  <NA>  nd56
#> 71  sp15  sp15
#> 72  <NA>  nd57
#> 73  <NA>  nd58
#> 74  <NA>  nd59
#> 75  <NA>  nd60
#> 76  <NA>  nd61
#> 77  <NA>  nd62
#> 78  <NA>  nd63
#> 79  <NA>  nd64
#> 80  <NA>  nd65
#> 81  sp16  sp16
#> 82  <NA>  nd66
#> 83  sp17  sp17
#> 84  sp18  sp18
#> 85  sp19  sp19
#> 86  <NA>  nd67
#> 87  <NA>  nd68
#> 88  <NA>  nd69
#> 89  <NA>  nd70
#> 91  <NA>  nd71
#> 92  <NA>  nd72
#> 93  <NA>  nd73
#> 94  sp21  sp21
#> 95  <NA>  nd74
#> 96  <NA>  nd75
#> 97  sp22  sp22
#> 98  sp23  sp23
#> 99  sp24  sp24
#> 100 <NA>  nd76
#> 101 <NA>  nd77
#> 102 sp25  sp25
#> 103 sp26  sp26
#> 104 <NA>  nd78
#> 105 sp27  sp27
#> 106 sp28  sp28
#> 107 sp29  sp29
#> 108 sp30  sp30
#> 109 sp31  sp31
#> 110 sp32  sp32
#> 111 sp33  sp33
#> 112 sp34  sp34
#> 113 sp35  sp35
#> 114 sp36  sp36
#> 115 sp37  sp37
#> 116 sp38  sp38
#> 117 sp39  sp39
#> 118 <NA>  nd79
#> 119 <NA>  nd80
#> 120 <NA>  nd81
#> 121 sp40  sp40
#> 122 sp41  sp41
#> 123 sp42  sp42
#> 124 <NA>  nd82
#> 125 sp43  sp43
#> 126 <NA>  nd83
#> 127 sp44  sp44
#> 128 sp45  sp45
#> 129 <NA>  nd84
#> 130 <NA>  nd85
#> 131 sp46  sp46
#> 132 sp47  sp47
#> 133 sp48  sp48
#> 134 <NA>  nd86
#> 135 <NA>  nd87
#> 136 sp49  sp49
#> 137 sp50  sp50
#> 139 sp52  sp52
#> 140 <NA>  nd88
#> 141 <NA>  nd89
#> 142 sp53  sp53
#> 143 sp54  sp54
#> 144 sp55  sp55
#> 146 sp57  sp57
#> 147 <NA>  nd90
#> 148 sp58  sp58
#> 149 <NA>  nd91
#> 150 sp59  sp59
#> 151 sp60  sp60
#> 152 sp61  sp61
#> 153 sp62  sp62
#> 154 sp63  sp63
#> 155 <NA>  nd92
#> 156 <NA>  nd93
#> 157 sp64  sp64
#> 158 sp65  sp65
#> 159 sp66  sp66
#> 160 <NA>  nd94
#> 161 <NA>  nd95
#> 162 <NA>  nd96
#> 163 <NA>  nd97
#> 164 sp67  sp67
#> 165 sp68  sp68
#> 166 sp69  sp69
#> 167 sp70  sp70
#> 168 sp71  sp71
#> 169 sp72  sp72
#> 170 sp73  sp73
#> 171 sp74  sp74
#> 172 <NA>  nd98
#> 173 sp75  sp75
#> 174 sp76  sp76
#> 176 sp78  sp78
#> 177 sp79  sp79
#> 178 sp80  sp80
#> 179 sp81  sp81
#> 180 sp82  sp82
#> 181 sp83  sp83
#> 182 sp84  sp84
#> 183 sp85  sp85
#> 184 sp86  sp86
#> 185 sp87  sp87
#> 186 sp88  sp88
#> 187 sp89  sp89
#> 188 sp90  sp90
#> 190 sp92  sp92
#> 191 sp93  sp93
#> 192 sp94  sp94
#> 193 sp95  sp95
#> 194 sp96  sp96
#> 195 sp97  sp97
#> 196 sp98  sp98
#> 197 sp99  sp99
#> 
#> $hist
#>    idx         t from to        x0         tc idx2 name2
#> 13  36 1.7712115    1  3 1.0455948 0.72561669    6   sp7
#> 7   54 1.4714900    1  3 1.2427894 0.22870062    9  sp10
#> 11  71 1.7523062    1  2 1.3112619 0.44104431   14  sp15
#> 6   81 1.3637622    1  3 1.3501228 0.01363937   15  sp16
#> 12 105 1.7655396    2  3 1.5479112 0.21762839   25  sp27
#> 9  109 1.6938940    2  3 1.5603822 0.13351183   29  sp31
#> 17 116 1.8579434    1  3 1.5923874 0.26555601   36  sp38
#> 20 125 1.9110279    1  3 1.6211681 0.28985983   41  sp43
#> 15 127 1.8226049    1  3 1.6225558 0.20004911   42  sp44
#> 10 137 1.7423445    3  1 1.6671139 0.07523055   48  sp50
#> 16 157 1.8506348    1  3 1.7573752 0.09325955   60  sp64
#> 19 158 1.9005004    2  3 1.7575846 0.14291580   61  sp65
#> 2    6 0.7937786    1  2 0.3155218 0.47825687   96   nd6
#> 14 161 1.7896296    1  2 1.7587350 0.03089453  124  nd95
#> 1   10 0.7664957    1  2 0.5968594 0.16963631  141  nd10
#> 5   43 1.2514209    1  3 1.1695460 0.08187492  148  nd35
#> 8  101 1.6656791    1  3 1.5298264 0.13585273  167  nd77
#> 4   12 1.1409286    1  3 0.6473286 0.49360002  181  nd12
#> 18 118 1.8697629    3  2 1.6061157 0.26364715  182  nd79
#> 3   27 1.1014641    1  3 0.9323727 0.16909142  184  nd24
#> 
#> $edge.state
#>   [1] 1 2 2 2 2 2 2 2 2 2 2 3 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 3 2 2 2 1
#>  [38] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 2 2 2 1 1 3 1 3 1 1 1 1 1 1
#>  [75] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 3 3 1 1 1 1 1 2 2 2 2 2 2 3 2 2 1
#> [112] 1 1 1 1 1 3 3 3 3 3 1 1 1 1 1 3 1 1 1 1 1 2 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [149] 1 3 3 3 3 3 3 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 3 1 1 1 1 1 3 2 2
#> [186] 2 3 3 3 1 1 3 3 3 3 1
#> 
#> $tip.state
#>  [1] "1" "2" "2" "1" "3" "2" "3" "2" "3" "2" "3" "2" "3"
#> 
#> attr(,"class")
#> [1] "phylo"
#> attr(,"order")
#> [1] "cladewise"