Simulate a birth/death/sampling tree

This tree can be passed to saasi, and will include speciation, extinction, sampling, and mutation events.

sim_bds_tree(
  params_df,
  q_matrix,
  x0,
  max_taxa = 100,
  max_t = 100,
  include_extinct = FALSE
)

Arguments

params_df

Data frame containing non-q parameters used in ancestral state reconstruction algorithm. Must have the following column names: state, prior, lambda, mu, and psi. The prior values refer to the baseline probabilities of the states (used at the root of the tree).

Example:

state	prior	lambda	mu	psi

1	1/3	3	0.02	1

2	1/3	3	0.02	1

3	1/3	3	0.02	1

q_matrix

Numeric q matrix used in ancestral state reconstruction algorithm. Row and column indices or names represent states.

x0

Natural number used as root state in the returned tree. Must be a state declared in params_df.

max_taxa

Maximum number of nodes allowed in the returned tree.

max_t

Maximum depth allowed in the returned tree.

include_extinct

Boolean declaring whether extinct taxa are included in the returned tree.

Value

A phylo phylogenetic tree (ape format).